requirements.txt
playwright~=1.52.0
pytest~=8.3.5
python-dotenv
allure-python-commons~=2.14.2